Ignore migrate versioning tables in utf8 sanity check
The migration version tables for alembic and sqlalchemy-migrate should
not be validated to be utf8. This is because these tools are not
explicitly part of the database schema defined in the projects. In certain
configurations it is not possible to ensure these tables will be utf8
character set unless a change to the my.cnf or explicit character set
on table creation is specified.

Change-Id: Ib5c0f93c788afda92aad862e50dc086dc39605f1
Closes-Bug: #1301036