Remove ValueError when accessing sys.modules
Previously there was ValueError that cannot be raised either by
__import__() nor getattr(). I think the original idea was to have KeyError
in order to catch exception when module is not part of sys.modules.

Since there is __import__ prior to getting class from sys.modules, there
is eliminated a chance of raising KeyError when getting key from
sys.modules. Either __import__ raises exception or import is successful
and module is present in sys.modules.

Change-Id: I1344a9e9d544b0c48e36f96be5b01911a7ec5497