Add funcName to audit/info message output
When trying to get a handle on a bug it is very frustrating to get a
log message that could have come from multiple places and be unable to
disambiguate them.  For example in [1] the last clue was an AUDIT
message that might have come from one of two places in the same file.

AUDIT and INFO level messages do give their name, but it doesn't help
when the duplication is with a file.  It seems reasonable to always
log the function name to avoid this.

As an example, there is one instance where this sort of thing happens
in service.py

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1084833

Change-Id: Ib59d77d94d9d0d5c0d2867c09f46784242633c19