Don't call provision.py directly
When testr.conf calls the sqlalchemy/provision.py script directly,
sys.path[0] is the directory that invoked the script.  So when it
tries to import openstack.common.db you end up with an import failure.

---
Traceback (most recent call last):
  File "./openstack/common/db/sqlalchemy/provision.py", line 27, in <module>
    from openstack.common.db import exception as exc
ImportError: No module named openstack.common.db
---

The solution is to run it with -m

Closes-Bug: #1303579
Change-Id: Ia7ff187df1685f0dd5ef32c3f6de474218874b04