Allow passing environment variables to execute()
Some of the projects might need to pass environment variables
to the processes they launch. For instance, if something needs
to rely on the process' output it should be able to pass LANG="C"
in order to set the correct localization.

This patch adds a env_variables parameter to processutils.execute()
which allows to pass news variables or replace the value of the existing
ones for the process.

Closes-bug: #1304284
Change-Id: If801e11391bf7cf66518533443c3b29af53653e5