Robust detection for postgresql connection errors
In most cases, when the database is failing over to a new node,
engine.dialect.is_disconnected() returns True. However in some
cases it returns False and the exception string is "unknown error".
In those cases we can however query the closed attribute on the
db_connection that is set to True at that point.

Change-Id: I64ef17f451780c622e1cfd257acf74e2e11e8b62
Closes-Bug: #1306668