Fix drop_unique_constraint utility for sqlite
Commit c76be5bbfe added the drop_unique_constraint utility method in
Havana but the way the method exists in oslo and how it exists (and
always existed in nova for the most part) aren't the same. Worse, the
method doesn't work for sqlite in migrations since sqlalchemy-migrate's
constraint dropper for sqlite doesn't support the 'alter table drop
constraint' call.

This change replaces the current drop_unique_constraint method in oslo
with the current version from nova.  There are three related unit tests
and they are nearly identical between nova and oslo, the only difference
being the nova version in one test was dropping the tables created in
the test and the oslo version wasn't, so I added that here.

This patch is also needed to not only make the method work with sqlite
but so that nova commit 1a8dd2487 can move forward with removing several
hundred lines of duplicate code from nova so nova can switch over to
using oslo's db utility methods.

Closes-Bug: #1307266
Related-Bug: #1280055

Change-Id: Iebdbed18d5d737aab46a2ad6b99c2d8c8ee4106a