Add API for creating translation functions
After gettextutils moves to a library, we will not be able to rely on
a single set of global translation functions. Each library will need
its own functions, using a different translation domain, so the right
message catalogs can be loaded.

Since we know when the factory is created which version of Python is
used and whether or not lazy translation should be enabled, return
functions that do the translation without checking those settings
at runtime for each message.

bp graduate-oslo-i18n

Change-Id: I0a177f70e71f1a5708d9e50ff0399eba6ec02c1c