Add a Retry decorator in excutils
The RetryDecorator gives the ability to retry a function upon
suggested exceptions.
The decorated function is retried for the given number of times, and the
sleep time between the retries is incremented until max sleep time
is reached. If the max retry count is set to -1, then the decorated
function is invoked indefinitely until an exception is thrown, and
the caught exception is not in the list of suggested exceptions.

This decorator will be useful in several projects (such as Glance,
oslo.vmware and Nova).

Change-Id: I535bf8281b733a15b6d76d8c7130fb51f89c5ab3