Move Microseconds to date-format instead of log-format
Microsecond resolution has been moved to the date-format string
instead of being embeded in the default log string. This makes it
possible to change the date format (including microseconds) without
needing to change both the log-formatting options and the
log-date-format option.

Moving the microsecond time resolution to the date formatter is done
by overriding the ``formatTime`` method and doing a simple string
sub on the %(msecs) substitution format string.

In the case that the legacy (deprecated) log-format option is used,
the %(msecs)03d is removed from the datefmt line (if it exists) as
the formatter doesn't know how to handle the substitution.

Change-Id: I653264cc556f547d4550bec72c88896f659ab854