Remove improper use of logging.getLevelName
The logging.getLevelName is supposed to be used to get string
representation from integer code, not vice versa. It is possible to get
int code from the string using this function in Python 2, but not in
Python 3. In any case, Logger.setLevel can accept string name of level
and thus there is no need to obtain integer code.

Change-Id: I4500681e842f0c1799931407fa66c660bdaf2d41