Remove improper use of logging.getLevelName
The logging.getLevelName is supposed to be used to get string
representation from integer code, not vice versa. It is possible to get
int code from the string using this function in Python 2, but not in
Python 3.4. Only in Python2.6, Logger.setLevel doesn't do automatic
conversion, and thus requires call to getLevelName.

Change-Id: I4500681e842f0c1799931407fa66c660bdaf2d41