Mask passwords that are included in commands
The current password masking doesn't scrub passwords from commands
that may be written to log files.  This commit adds support for
scrubbing passwords provided as options with commands, where the
tokens in the command are separated by commas and wrapped in quotes.

Adds tests to ensure commands are properly sanitized.

Change-Id: I37b9a80142ec5dcadb731332d8c5f494bdc7bfc1
Closes-Bug: #1320028