Mask passwords that are included in commands
The current password masking doesn't scrub passwords from commands
that may be written to log files.  This commit adds support for
scrubbing passwords provided as options with commands, where the
tokens in the command are separated only by spaces.  There is a
separate commit for scrubbing passwords when the tokens are separated
by commas and wrapped in quotes.

Adds tests to ensure commands are properly sanitized.

Change-Id: Idc8d55d5ebcc4fc5a3f66efeff86b19c3d94e7e2
Closes-Bug: #1320712