Make script to run in error msg configurable
This commit adds new variables to the check_uptodate.sh script to set
the error message if the sample is out of date. One of the new
variables SCRIPT_PATH can be set as an env variable to specify
alternate scripts to run to reset the config sample.

This is needed by Tempest for example because it wraps the oslo
version of generate_sample.sh to set env variables needed to load all
the config options.

Change-Id: Ibdeaf981364b3e63f5e23c2de2c3717821f0ee8f