Use correct "coding" file variable to set file encoding
Like Python, Emacs can read the file encoding from the first two
lines. Python will accept a line with

  -*- encoding: <some-encoding> -*-

since it looks for containing "coding:" or "coding=" (PEP 263) and
allows substring matches. Emacs only accepts

  -*- coding: <some-encoding> -*-

since it looks for a file variable named "coding", not "encoding":

  https://www.gnu.org/software/emacs/manual/html_node/emacs/Specify-Coding.html

Setting a variable called "encoding" has at best no effect, but it can
result in Emacs prompting the user with

  The local variable list in __init__.py contains values that may not
  be safe.

when opening the file. This is because the "encoding" variable is not
whitelisted as a safe variable.

The command used to update the files were:

  sed -i '1,2s/-\*- encoding:/-*- coding:/' **/*.py

This accounts for the case where the file variables is set in the
second line because the first line is used for a "#!" line.

Change-Id: I2d7be303b515be0e3e891f041477c99cedcca1bb