Make memory cache eviction use priority queue
The entire memory cache was being scanned for items which had passed
their timeout every time an item was fetched.  This could cause a
performance problem for large caches.

This was fixed using a priority queue to keep track of the timeouts. The
youngest timeouts will be at the top of the heap and can be easily
evicted from the cache, without scanning the entire cache.

Change-Id: Id49e21ffd70862d6abd1b09c0cde801100c9a562
Closes-Bug: #1276530