Mask passwords included without quotes at the ends of commands
The current password masking doesn't scrub passwords from commands
in the case where the password doesn't have quotes around it and
the password is the last string in the command.  This commit updates
one of the regular expressions to catch this case.

Adds tests to ensure passwords at the ends of commands are properly
sanitized.

Change-Id: Id57a0cb05cd76ef8c26def738305ade6b085aaa7