cfgfilter: don't re-register options
If the underlying config object already has a given option registered
we shouldn't re-register it with the filter, but instead just treat
it like an imported option.

The use case here is unit tests which register options and set
overrides for them before passing the config object to the code which
registers the options with a filter wrapper.

Change-Id: I50791f1b47083626babb79ec29e503913f7edc2f