Add guard in case human_id target attribute is None
`human_id` is a property to facilitate bash-completion by providing a
consistent way for users to address an object in human-readable way.

Some objects, like Glance image names, however, are nullable which can cause
the slugify code to raise an exception.

The solution is to just add a guard and avoid slugifying the null field if we
detect that case.

Change-Id: Id8e3535b121f916a9814ec727185c2fe5554feb0
Closes-Bug: 1328684