Prevent races in opportunistic db test cases
Opportunistic db test cases use openstack_citest database credentials
to create schemas on demand in tests. This actually can conflict with
existing migrations tests in Nova and possibly other projects which
drop and recreate openstack_citest database, so that if there was an
open connection to openstack_citest database in an opportunistic test
case, Nova migrations test would fail trying to drop the database.

The solution here is to use openstack_citest user credentials but not
connect to openstack_citest database when creating/dropping schemas:
for MySQL we can use an empty db name and for PostgreSQL the service
postgres database.

Closes-Bug: #1328997

Change-Id: Id82090743e8b93f6c960de066fd05d8a2f9fb934