Make periodic tasks run on regular spacing interval
Instead of setting the value of _periodic_last_run[task_name] to be the
current time it is set as the nearest multiple of the spacing interval
which is in the past with added jitter. This will ensure the task runs
regularly but avoids synchronizing that interval with other nodes.

This patch also removes the coalescing of tasks which are _nearby_ (0.2
seconds away).

Change-Id: I118ce960c2a7f53130fb4240de3e1574b7e06e63
Closes-Bug: #1326020