Changes calcuation of variable delay
In the loopingcall.FixedIntervalLoopingCall class, an warning is logged
when 'delay' is equal or less than zero. Commonly delayed time is
considered something over due which is a positive value. Also a task
delayed zero seconds is not considered delayed but finished on time.
Therefore this patch fixes the calculation of the variable 'delay' and
omits a warning log when 'delay' is zero

Change-Id: I9dbae0838e6c03fa0eaa79e7332e650b02ed9c77