Mask any password in exception command string
Currently the command executed by processutils.execute
is included in the exception (if one is generated). This
could contain passwords and other information that could
represent a security vulnerability.

Also added a test for this, ensuring that stdout and
stderr contain the expected text, and that the password
is obscured from the command that is part of the exception.

OSSO is aware of this change.

Change-Id: I2912ad7efc10b065c954fd198325305f81a8c6ec
Closes-Bug: #1343604
SecurityImpact: