Mask any password in exception command string
Currently the command executed by processutils.execute
is included in the exception (if one is generated). This
could contain passwords and other information that could
represent a security vulnerability.

Also added a test for this, ensuring that stdout and
stderr contain the expected text, and that the password
is obscured from the command that is part of the exception.

Before merging this change into other projects, it is
important to understand that the command being returned
in a ProcessExecutionError exception (should one be
thrown) is one where passwords will have been masked
using strutils.mask_password(). Therefore the command
may no longer be executable. This may break some existing
code that may have a construct such as:

     except ProcessExecutionError as err:
     	    # retry the command
     	    processutils.execute(err.cmd)

OSSO is aware of this change.

DocImpact:
The documentation for the execute command should include
a note about the change in behavior described in the message
as above.

SecurityImpact:

Change-Id: I2912ad7efc10b065c954fd198325305f81a8c6ec
Closes-Bug: #1343604