Log information when processutils.execute encounters an error
It would help debugging a problem if more information, such as stdout
and stderr are automatically logged when an error occurs in executing
a command using processutils.execute().

Currently we only throw an exception and leave logging to the caller.
This is resulting in a lot of code just logging the exception when it
could just as well be done right here by execute(). This would
also eliminate code such as the one proposed in the fix for 1340124.

We only log the error when the user has chosen not to
ignore_exit_code and when the exit_code is not in the ones which
the user has asked to ignore (check_exit_code).

This adds a new parameter to execute() called 'verbose'.

This problem was originally reported by Ionuț Arțăriși in bug
1340124. Peter Stachowski proposed some aspects of this fix
in Trove. Both of them are credited below.

Closes-Bug: #1342857
Suggested-By: Peter Stachowski <peter@tesora.com>
Reported-By: Ionuț Arțăriși <iartarisi@suse.cz>

Doc-Impact:
New parameter added to processutils.execute().

Change-Id: Ied5a732b8845c5ea9f79727cb985711d324d834f