Mask passwords in exceptions and error messages; improve logging
This patch set fixes two bugs, 1343604 and 1342857.

When a ProcessExecutionError is thrown by processutils.execute(), the
exception may contain information such as password. This is bug
1343604. Upstream applications that just log the message (as several
appear to do) could inadvertently expose these passwords to a user
with read access to the log files. It is therefore considered prudent
to invoke strutils.mask_password() on the command, stdout and stderr
in the exception. A test case has been added to ensure that all three
are properly masked.

In addition, it would help debugging considerably if a failure in
execute() generated a message (at a client specified logging
level). This is bug 1342857. Such a message, generated by execute()
would only occur if the user requested it by specifying a parameter
(verbose), and execute() were to throw an exception. This message that
gets logged would include properly sanitized information. As an
example, this change would address the issue being addressed in bug
1340124 for the Trove project only.

These two bugs collectively reduce the risk of accidentally leaking
information such as passwords in exceptions and in log files. They
will also simplify the code in clients which do not have to go through
the motions of sanitization and logging and can instead just get the
messages logged at the source.

OSSA is aware of this change request.

Change-Id: I173dfb865e84eb7dee54a22c76db1e4f125a0a8a
Closes-Bug: #1343604
Closes-Bug: #1342857
Doc-Impact: New parameter added to processutils.execute().
Suggested-By: Peter Stachowski <peter@tesora.com>
Reported-By: Ionuț Arțăriși <iartarisi@suse.cz>