Mask passwords in exceptions and error messages
When a ProcessExecutionError is thrown by processutils.execute(), the
exception may contain information such as password. This is bug
1343604. Upstream applications that just log the message (as several
appear to do) could inadvertently expose these passwords to a user
with read access to the log files. It is therefore considered prudent
to invoke strutils.mask_password() on the command, stdout and stderr
in the exception. A test case has been added to ensure that all three
are properly masked.

OSSA is aware of this change request.

Change-Id: I173dfb865e84eb7dee54a22c76db1e4f125a0a8a
Closes-Bug: #1343604