Log stdout, stderr and command on execute() error
It would help debugging considerably if a failure in execute()
generated a message (at a client specified logging level). Such a
message, generated by execute() would only occur if the user requested
it by specifying a parameter (verbose), and execute() were to throw an
exception. As an example, this change would address the issue being
addressed in bug 1340124 for the Trove project only.

Change-Id: If8180e042e1a05374ef705f5873cc0981243619f
Closes-Bug: #1342857
Doc-Impact: New parameter added to processutils.execute().
Suggested-By: Peter Stachowski <peter@tesora.com>
Reported-By: Ionuț Arțăriși <iartarisi@suse.cz>