Switch logging debug on by default
Our defaults should match what our users use and we've had fairly
comprehensive feedback that operators - particularly large ones - use
debug logging to facilitate first-failure-analysis of faults.
Secondly, all devstack installs force debug to True, (and we depend on
that in CI) - changing this to match how these environments are
deployed will allow us to run defaults in a wider set of environments.
In future when our logging is fixed up we may find that deployers no
longer want debug, and we can of course change our defaults then to
follow suite.

I'm told there are some particularly egregious modules that need to be
filtered for this to make sense - I'm going to request operator input
to get those into this patch before it lands.

Change-Id: I54a694e55bba80bf4818486246ab12ea132ee5bd