Switch oslo-incubator to use oslo.utils
keystoneclient, keystonemiddleware, nova all have tests that mock
utcnow when we introduce oslo.utils into those projects, we have
test failures because of references to copies of timeutils/utcnow
in for example memorycache. So the best way to deal with it is to
ensure that all code in oslo-incubator refer to oslo.utils.

Closes-Bug: #1354521
Change-Id: If80e66c7dae1bfcda8e884c57b7086f34d026040