Stop substituting jsonutils as a anyjson reference impl
I've been able to reliably cause kombu to fail when kombu
starts to use the jsonutils implementation that is forced
onto it whenever jsonutils is loaded.

An example: http://paste.openstack.org/show/102226/

Since jsonutils is copied across many many different projects
and libraries it does not seem safe or smart to have that file
each time it is imported to force anyjson that this newly
loaded module is the de-facto json impl. This is especially
problematic since jsonutils loads and dumps do specific things
that are mainly openstack specific and now those openstack
specific adjustments become forced onto any library that also
uses anyjson...

This is especially problematic now that jsonutils has
conditionals like the following in dumps():

if is_simplejson:
    kwargs['namedtuple_as_object'] = False

Now all other anyjson using modules or libraries are affected by
this decision to do this, whether they like that or not...

Change-Id: I4faf2e4ef64dd5f666dfe42bd5f26b12b3daa73e