Stop substituting jsonutils as a anyjson reference impl
I've been able to reliably cause kombu to fail when kombu
starts to use the jsonutils implementation that is forced
onto it whenever jsonutils is loaded.

An example: http://paste.openstack.org/show/102226/

Since jsonutils is copied across many many different projects
and libraries it does not seem safe or smart to have that file
each time it is imported to force anyjson that this newly
loaded module is the de-facto json impl.

Change-Id: I4faf2e4ef64dd5f666dfe42bd5f26b12b3daa73e