Don't write python bytecode while testing
This should reduce the number of times you need to do things like git
clean -x -i to kill pyc files in your local directory to get tests to
pass.
It doesn't appear to address pyc files in the .tox directories, and
attempts to address that seem to go in circles. Ideas welcomed.

Change-Id: Ic0f74c02b94f40723364055316e8dd2e7dffa8ad
Related-Bug: #1368661