Fix using multiple instances of ProcessLauncher
Make pipe and signal handler setting a class domain functionalities that
are done only once, and instruct the instances to only manage it's own
spawned children processes and to not assume that all sub processes are
owned by each instance.

Unit tests was refactored in this patch to make better reuse of code.

Change-Id: Ica4c38bd3ef31292574d671a87632a1459031b57
Closes-Bug: 1364876