Remove oslo.log code and clean up versionutils API
Remove the modules and tests that have moved to oslo.log. Replace the
use of oslo.log with the standard library logging module.

Replace calls to LOG.deprecated() with a new function in versionutils
that is copied from the body of the old implementation.

Add fatal_deprecations configuration option to versionutils, where it is
used now instead of the old log module.

Change-Id: I5903f9a9c521789cbf3381d9741a8a5308767e2d
Closes-Bug: #1373120
Blueprint: fix-import-cycle-log-and-versionutils