Add middleware.request_id shim for Kilo
A deprecation warning was added to oslo-incubator for Juno, but
unfortunately not to any of the projects actually using it.

Since this module is referenced in the api paste config, removing
the module from the projec trees without deprecation warning
constitutes an operator-visible backwards incomatible change,
which AIUI is not allowed.

This mitigates the lack of deprecation period by adding a shim for
Kilo, which projects can sync before switching their paste defaults
to point to oslo.middlware, we can then switch the projects paste
configs to point at oslo.middleware, and remove the shim in L.

Change-Id: I11caf1119c44712ed1d15810703e38c50e009709
Related-Bug: #1380629