Traverse children list to check if any worker stop
Currently Neutron server spawns two green threads for api and rpc
job. In _wait_child, both green threads get pid of the stopped
worker from os.waitpid, but if the pid is not in the green thread's
children list, the stopped worker is not restarted.

Considering this two green threads are in the same process, api
green thread may get a pid belonging to a worker forked by rpc
green thread and will not restart the worker. We need to traverse
the children list of each green thread in _wait_child to ensure
stopped worker can be properly recognized.

Change-Id: I056235b11f3e239a8a7e78a507e002978a306b8c
Closes-Bug: #1384580