Fixed a problem with neutron http policy check
When a rule "create_network": "http://127.0.0.1:8080/ or rule:default"
is configured in policy.json, create network with default values reports
a "Circular reference detected" exception.
The reason is, for unspecified values, default value of
neutron.api.v2.attributes.ATTR_NOT_SPECIFIED (that is, object())
is assigned by Neutron.
This creates a problem when jsonutils.dumps() is executed on a JSON
string that contains object(), thus reporting the said exception.
The fix replaces object() into empty string value {} in the
dictionary only for the policy check.

Change-Id: Ibd4859ed8c919ddb32cb75cc58c7f91d0e84a458
Closes-Bug: 1233707