cliutils restore python3 string decode
Since the the safe_encode function moved from incubator strutils
to oslo.utils the print_list and print_dict functions don't work
as expected on python3, because the safe_encode in encodeutils
lacks the logic to decode on python3 like existed in strutils.

So, reinstate that logic within the print_list and print_dict
funtions, and add a couple of simple tests to ensure we keep
them working (no other tests check the string output from the
prints AFAICS).

Change-Id: I42f6aa802c1f94fcbf0e9170024cf37089f92782
Closes-Bug: #1390505