Redis: for non-strict Redis class, expired ttl is None, not -1
Python bindings for Redis return None if ttl is expired instead of -1
when Redis class is used (instead of StrictRedis) which is the default
for oslo-incubator. Fixed unit test to pass the check.

The patch does not belong to master where old RPC layer was removed
since Juno. Neither it belongs to oslo.messaging since the test was not
copied there.

Closes-Bug: #1259532
Change-Id: Id54e78ca56bb5a5bf1cdad31bc5965ff02c0d9b4