Improve unreleased change report
Add a script to sort the version tags so that list_unreleased_changes.sh
does not pick up "eol" tags and other values that shouldn't be
considered when determining the most recent tag.

Also add more detail to the output for each library, including diff
stats and the actual diffs for requirements files.

Change-Id: Id091d697810ba7d65ae64fd2e52fd09cd78baf59