Port processutils to Python 3
ProcessExecutionError now decodes stdout and stderr to build the
exception message. It avoids to see "stdout: b'...'" in the message (you
get "stdout: '...'" instead).

Fix also unit tests to use bytes strings for stdin, stdout and stderr.

The mask_password() function is wrapped to return bytes strings when the
input message is a bytes string.

Change-Id: I16970ddf3f78fc844d8795c6a9428c7b9c23ae59